requirements.txt
streamlit
python-dotenv
google-generativeai>=0.5.0
PyPDF2
langchain>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.0.20
faiss-cpu